Versão final

+++ src/lib/pages/arquivo_configuracao.dart
@@ -6,81 +6,71 @@ import 'package:path_provider/path_provider.dart';
 
 Map<String, dynamic> listaConfiguracoes;
 
-// Map<String, dynamic> listaConfiguracoesPadrao = {
-//   "tamanhoFonte": 30.0,
-//   "corFonte": (Colors.black).value,
-//   "corTema": (Colors.blueAccent[400]).value,
-//   "usuarioEscolheu": false
-// };
-
 Map<String, dynamic> listaConfiguracoesPadrao() {
-  // Map<String, dynamic> lalal =
   return {
-  "tamanhoFonte": 30.0,
-  "corFonte": (Colors.black).value,
-  "corTema": (Colors.blueAccent[400]).value,
-  "usuarioEscolheu": false};
-
-  // return lalal;
+    "tamanhoFonte": 30.0,
+    "corFonte": (Colors.black).value,
+    "corTema": (Colors.blueAccent[400]).value,
+    "usuarioEscolheu": false
+  };
 }
 
 Color fonteNecessaria() {
   if (!listaConfiguracoes['usuarioEscolheu'])
-    return useWhiteForeground(Color(listaConfiguracoes['corTema'])) ? Colors.white : Colors.black;
+    return useWhiteForeground(Color(listaConfiguracoes['corTema']))
+        ? Colors.white
+        : Colors.black;
   else
     return Color(listaConfiguracoes['corFonte']);
 }
 
-Future restaurarPadrao() async{
-  final diretorio = await getApplicationDocumentsDirectory();    
+Future restaurarPadrao() async {
+  final diretorio = await getApplicationDocumentsDirectory();
   bool arquivoExiste;
   arquivoExiste = false;
   arquivoExiste = await File("${diretorio.path}/data.json").exists();
 
-  if (arquivoExiste)
-    await excluirArquivo();
+  if (arquivoExiste) await excluirArquivo();
 
   listaConfiguracoes = listaConfiguracoesPadrao();
   print('chegou aqui');
 }
 
-Future excluirArquivo() async{
-  final diretorio = await getApplicationDocumentsDirectory();    
+Future excluirArquivo() async {
+  final diretorio = await getApplicationDocumentsDirectory();
   await File("${diretorio.path}/data.json").delete();
 }
 
-Future<File> retornaArquivo() async{
+Future<File> retornaArquivo() async {
   final diretorio = await getApplicationDocumentsDirectory();
   return File("${diretorio.path}/data.json");
 }
 
-Future<String> lerDadosArquivo() async{
-  try{
+Future<String> lerDadosArquivo() async {
+  try {
     final file = await retornaArquivo();
     return file.readAsString();
-  }
-  catch (e){
+  } catch (e) {
     return null;
   }
 }
 
-Future saveData(dynamic arquivoDeConfiguracao) async{
-  String data = json.encode(arquivoDeConfiguracao);    
+Future saveData(dynamic arquivoDeConfiguracao) async {
+  String data = json.encode(arquivoDeConfiguracao);
 
   final file = await retornaArquivo();
   return file.writeAsString(data);
 }
 
 Future lerArquivo() async {
-    await lerDadosArquivo().then((data){      
-      listaConfiguracoes = json.decode(data);      
-    });
+  await lerDadosArquivo().then((data) {
+    listaConfiguracoes = json.decode(data);
+  });
 }
 
 Future abrirArquivo() async {
   final diretorio = await getApplicationDocumentsDirectory();
   bool arquivoExiste = await File("${diretorio.path}/data.json").exists();
 
-  if (!arquivoExiste)
-    saveData(listaConfiguracoesPadrao());
+  if (!arquivoExiste) saveData(listaConfiguracoesPadrao());
 }
